Big refactor and changed cheat to unlock button.
