+++ assets/imsim.css
@@ -1,5 +1,6 @@
 @import url("https://fonts.googleapis.com/css2?family=IBM+Plex+Sans:wght@400;500;600;700&family=Space+Grotesk:wght@500;700&display=swap");
 
+/* Tokens and theme */
 :root {
   --imsim-bg-top: #f7f8f5;
   --imsim-bg: #eef1ec;
@@ -34,7 +35,11 @@
   --imsim-size-dashboard-panel-max: 15rem;
   --imsim-size-lesson-grid-min: 38rem;
   --imsim-size-lesson-grid-max: 42rem;
-  --imsim-lesson-gap: 0.75rem;
+  --imsim-space-panel-gap: 0.75rem;
+  --imsim-space-card-gap: 0.85rem;
+  --imsim-space-panel-pad: 0.95rem;
+  --imsim-space-panel-pad-compact: 0.85rem;
+  --imsim-lesson-gap: var(--imsim-space-panel-gap);
   --imsim-lesson-card-pad: 0.9rem;
   --imsim-lesson-card-gap: 0.55rem;
   --imsim-slider-rail-size: 0.375rem;
@@ -66,6 +71,7 @@
   --imsim-table-hover: rgba(49, 95, 143, 0.055);
 }
 
+/* Base shell */
 body {
   font-family: "IBM Plex Sans", "Segoe UI", sans-serif;
   line-height: 1.5;
@@ -199,7 +205,8 @@ body {
 }
 
 .academy-card {
-  height: auto;
+  width: 100%;
+  height: 100%;
 }
 
 .academy-card.shell-card {
@@ -207,14 +214,22 @@ body {
 }
 
 .academy-card .card-body {
-  display: grid;
-  grid-template-rows: auto;
-  align-content: start;
+  display: flex;
+  flex-direction: column;
   gap: 0.55rem;
+  height: 100%;
+}
+
+.academy-card-content {
+  display: flex;
+  flex-direction: column;
+  flex: 1 1 auto;
+  gap: 0.55rem;
+  min-height: 100%;
 }
 
 .academy-card .button-block {
-  margin-top: 0.35rem !important;
+  margin-top: auto !important;
 }
 
 .academy-card-kicker,
@@ -275,27 +290,6 @@ body {
 .academy-status-badge {
   justify-self: start;
   margin-top: 0.1rem;
-}
-
-.cheat-code-row {
-  display: flex;
-  justify-content: flex-end;
-  min-height: 1.25rem;
-  padding-top: 0.45rem;
-}
-
-.cheat-code-button {
-  color: var(--imsim-muted);
-  font-size: 0.9rem;
-  line-height: 1;
-  opacity: 0.18;
-  padding: 0 0.2rem;
-}
-
-.cheat-code-button:hover,
-.cheat-code-button:focus-visible {
-  color: var(--imsim-blue);
-  opacity: 1;
 }
 
 .academy-result-alert {
@@ -603,6 +597,7 @@ body {
   padding-bottom: 0;
 }
 
+/* Structural cards and panels */
 .shell-card,
 .metric-card,
 .footer-card {
@@ -615,6 +610,18 @@ body {
 .dashboard-shell .shell-card,
 .lesson-intro-grid .shell-card {
   box-shadow: none;
+}
+
+.dashboard-panel {
+  min-width: 0;
+}
+
+.dashboard-panel-card {
+  width: 100%;
+}
+
+.panel-body-compact {
+  padding: var(--imsim-space-panel-pad-compact);
 }
 
 @supports not ((-webkit-backdrop-filter: blur(1px)) or (backdrop-filter: blur(1px))) {
@@ -692,6 +699,7 @@ body {
   color: var(--imsim-muted);
 }
 
+/* Controls and buttons */
 .control-field {
   display: grid;
   gap: 0.45rem;
@@ -969,6 +977,7 @@ body {
   transform: none;
 }
 
+/* Metrics, uploads, tables, and grids */
 .speed-readout {
   display: inline-flex;
   align-items: center;
@@ -1377,6 +1386,7 @@ body {
   color: var(--imsim-muted);
 }
 
+/* Sliders and chart surfaces */
 .rc-slider-rail {
   background-color: rgba(83, 98, 119, 0.22);
   height: var(--imsim-slider-rail-size);
@@ -1468,6 +1478,7 @@ body {
   width: 100%;
 }
 
+/* Dashboard layouts */
 .dashboard-shell #graph-panel > .card,
 .dashboard-shell #graph-panel > .card > .card-body {
   height: auto;
@@ -1866,6 +1877,7 @@ body {
   overflow: visible;
 }
 
+/* Responsive adjustments */
 @media (min-width: 1200px) {
   .control-rail-column {
     max-width: var(--imsim-size-control-rail-max);
