+++ src/imsim/ui/components.py
@@ -1,2154 +1,73 @@
 from __future__ import annotations
 
-from dataclasses import asdict
-
-import dash_ag_grid as dag
-import dash_bootstrap_components as dbc
-import pandas as pd
-import plotly.graph_objects as go
-from dash import Patch, html
-
-from ..models import GlobalSettings, InventoryItem, SimulationState
-from ..services.planning import (
-    effective_review_cycle,
-    format_money,
-    item_on_order,
-    update_gs_related_values,
+from .academy_components import (
+    academy_level_card_children,
+    academy_progress_children,
+    academy_result_children,
+    lesson_compact_summary_children,
+    lesson_locked_children,
+    lesson_objective_children,
+    lesson_tutorial_children,
+    reference_modal_children,
+    simulator_unlock_children,
 )
-from ..services.training import (
-    academy_level_status,
-    academy_levels,
-    active_layout_variant,
-    active_level,
-    evaluate_active_lesson,
-    fill_rate,
-    lesson_days_remaining,
-    visible_columns,
+from .figures import (
+    _plot_base_layout,
+    _plot_line,
+    _plot_marker,
+    _plot_marker_outline,
+    build_inventory_figure,
+    inventory_graph_style,
+    refresh_inventory_figure,
 )
-
-
-def _items_frame(items: list[InventoryItem]) -> pd.DataFrame:
-    if not items:
-        return pd.DataFrame()
-    return pd.DataFrame([item.to_dict() for item in items]).reset_index(names="idx")
-
-
-def _figure_theme(theme: str) -> dict[str, str]:
-    if theme == "dark":
-        return {
-            "plot_bg": "rgba(10, 19, 32, 0.72)",
-            "surface": "#121f31",
-            "text": "#e6eefc",
-            "muted": "#9aabc6",
-            "line": "rgba(214, 228, 255, 0.14)",
-            "guide": "rgba(148, 163, 184, 0.7)",
-            "pna": "#2dd4bf",
-            "proposed": "#d9904a",
-            "ats": "#60a5fa",
-            "zero": "#f87171",
-        }
-    return {
-        "plot_bg": "rgba(252,253,249,0.86)",
-        "surface": "#fcfdf9",
-        "text": "#15231f",
-        "muted": "#5b6964",
-        "line": "rgba(21, 35, 31, 0.13)",
-        "guide": "rgba(21, 35, 31, 0.5)",
-        "pna": "#167264",
-        "proposed": "#b9621e",
-        "ats": "#315f8f",
-        "zero": "#b33b3b",
-    }
-
-
-_ROOT_FONT_SIZE_PX = 16
-_PLOT_TITLE_SIZE_REM = 1.5
-_PLOT_MARGIN_REM = {"l": 1.5, "r": 1.5, "t": 3.5, "b": 1.5}
-_PLOT_MARGIN_COMPACT_REM = {"l": 1.5, "r": 1.5, "t": 1.5, "b": 2.4}
-_PLOT_LINE_WIDTH_REM = 0.1875
-_PLOT_MARKER_OUTLINE_WIDTH_REM = 0.125
-_PLOT_MARKER_SIZE_REM = {
-    "lesson": 0.5,
-    "lesson_backorder": 0.4375,
-    "signal": 0.875,
-    "signal_proposed": 0.625,
-    "signal_ats": 0.5625,
-    "signal_zero": 0.5,
-    "signal_stockout": 1.125,
-}
-_LESSON_ONE_GRAPH_HEIGHT = 340
-_LESSON_TWO_GRAPH_HEIGHT = 392
-_FILL_RATE_GRAPH_HEIGHT = 400
-
-
-def _rem_to_px(rem: float) -> float:
-    return rem * _ROOT_FONT_SIZE_PX
-
-
-def _plot_base_layout(
-    title: str | None,
-    colors: dict[str, str],
-    *,
-    include_margin: bool = True,
-    height: int | None = None,
-) -> dict[str, object]:
-    muted = colors.get("muted", colors["text"])
-    layout: dict[str, object] = {
-        "paper_bgcolor": "rgba(0,0,0,0)",
-        "plot_bgcolor": colors["plot_bg"],
-        "font": {
-            "color": colors["text"],
-            "family": "IBM Plex Sans, Segoe UI, sans-serif",
-            "size": 13,
-        },
-        "legend_title_text": "",
-        "legend": {
-            "font": {"color": muted, "size": 12},
-            "orientation": "h",
-            "yanchor": "bottom",
-            "y": 1.02,
-            "xanchor": "right",
-            "x": 1,
-        },
-        "hoverlabel": {
-            "bgcolor": colors["surface"],
-            "bordercolor": colors["line"],
-            "font": {"color": colors["text"]},
-        },
-        "xaxis": {
-            "gridcolor": colors["line"],
-            "linecolor": colors["line"],
-            "tickfont": {"color": muted, "size": 11},
-            "zerolinecolor": colors["line"],
-        },
-        "yaxis": {
-            "gridcolor": colors["line"],
-            "linecolor": colors["line"],
-            "tickfont": {"color": muted, "size": 11},
-            "zerolinecolor": colors["line"],
-        },
-    }
-    if title:
-        layout["title"] = {
-            "text": title,
-            "font": {"color": colors["text"], "size": _rem_to_px(_PLOT_TITLE_SIZE_REM)},
-        }
-    if include_margin:
-        margin_source = _PLOT_MARGIN_REM if title else _PLOT_MARGIN_COMPACT_REM
-        layout["margin"] = {side: _rem_to_px(size) for side, size in margin_source.items()}
-    if height is not None:
-        layout["height"] = height
-    return layout
-
-
-def _plot_line(color: str, *, dash: str | None = None) -> dict[str, object]:
-    line: dict[str, object] = {"width": _rem_to_px(_PLOT_LINE_WIDTH_REM), "color": color}
-    if dash:
-        line["dash"] = dash
-    return line
-
-
-def _plot_marker_outline(color: str) -> dict[str, object]:
-    return {"width": _rem_to_px(_PLOT_MARKER_OUTLINE_WIDTH_REM), "color": color}
-
-
-def _plot_marker(
-    color: str,
-    size_rem: float,
-    *,
-    symbol: str | None = None,
-    line: dict[str, object] | None = None,
-) -> dict[str, object]:
-    marker: dict[str, object] = {"size": _rem_to_px(size_rem), "color": color}
-    if symbol:
-        marker["symbol"] = symbol
-    if line:
-        marker["line"] = line
-    return marker
-
-
-def _lesson_snapshot_table(
-    label: str, headers: tuple[str, ...], rows: tuple[tuple[str, ...], ...]
-) -> html.Div:
-    return html.Div(
-        [
-            html.Div(label, className="lesson-snapshot-label"),
-            html.Table(
-                [
-                    html.Thead(html.Tr([html.Th(header) for header in headers])),
-                    html.Tbody([html.Tr([html.Td(value) for value in row]) for row in rows]),
-                ],
-                className="lesson-service-table",
-            ),
-        ],
-        className="lesson-snapshot-block",
-    )
-
-
-def _lesson_snapshot_disclosure(
-    body: html.Div,
-    *,
-    label: str = "Show snapshot details",
-    hint: str = "Status and goal details",
-    open_by_default: bool = False,
-) -> html.Details:
-    props: dict[str, object] = {"className": "lesson-snapshot-disclosure"}
-    if open_by_default:
-        props["open"] = True
-    return html.Details(
-        [
-            html.Summary(
-                [
-                    html.Span(label),
-                    html.Span(hint, className="lesson-disclosure-hint"),
-                ],
-                className="lesson-disclosure-summary",
-            ),
-            html.Div(body, className="lesson-disclosure-body"),
-        ],
-        **props,
-    )
-
-
-_GLOSSARY_ENTRIES = (
-    (
-        "PNA",
-        "Projected net available. The inventory position used for replenishment decisions. "
-        "Full training formula: on hand - reserved - committed - backordered + on "
-        "order + received.",
-    ),
-    (
-        "On hand",
-        "Physical stock currently available in the warehouse before considering future "
-        "receipts or unresolved demand.",
-    ),
-    ("On order", "Quantity already ordered but not yet received."),
-    ("Backorder", "Demand that could not be filled from available stock."),
-    ("Usage rate", "The monthly demand pace used by the simulator."),
-    ("Daily usage", "Usage rate divided by day basis."),
-    ("Lead time", "The time between placing an order and receiving it."),
-    ("OP", "Order point. The level where replenishment should begin."),
-    (
-        "Safety stock",
-        "Extra stock added to protect against normal demand or lead-time variation.",
-    ),
-    (
-        "LP",
-        "Line point. The higher review threshold used to decide what else should ride "
-        "along during a line review.",
-    ),
-    ("OQ", "Order quantity. The practical quantity to buy when replenishment is needed."),
-    (
-        "EOQ",
-        "Economic order quantity. A calculated quantity that balances replenishment cost "
-        "and carrying cost.",
-    ),
-    (
-        "SOQ",
-        "Suggested order quantity. The simulator's recommendation after considering PNA, "
-        "OP/LP, OQ, and standard pack rounding.",
-    ),
-    ("Standard pack", "Supplier pack increment used to round suggested buys."),
-    ("Critical point", "Urgent replenishment threshold below normal protection."),
-    (
-        "Surplus threshold",
-        "LP + OQ in the simulator's short-term overstock model.",
-    ),
-    (
-        "Fill rate",
-        "Percent of stock lines completed without a stockout miss.",
-    ),
-    (
-        "Hits",
-        "How often an item appears on demand lines. Used to show customer frequency/popularity.",
-    ),
-    (
-        "ASQ",
-        "Average sales quantity. Keep locked until advanced and certification content.",
-    ),
+from .grids import (
+    _grid_theme_class,
+    build_custom_order_grid,
+    build_custom_order_rows,
+    build_inventory_table,
+    build_po_overview_grid,
+    build_po_overview_rows,
+    custom_order_grid_options,
+    po_overview_grid_options,
+)
+from .panels import (
+    build_exception_center,
+    build_kpi_strip,
+    costs_card_children,
+    github_footer_card,
+    sales_card_children,
+    service_card_children,
 )
 
-
-def _format_pct(value: float | None) -> str:
-    return "—" if value is None else f"{value:.1f}%"
-
-
-def _compact_lesson_pill_text(row: str) -> str:
-    replacements = (
-        ("Current fill rate: ", "Fill rate: "),
-        ("Current after-overhead GM: ", "After-OH GM: "),
-        ("After-overhead GM: ", "After-OH GM: "),
-        ("Items closing at/above OP: ", "At/above OP: "),
-        ("Close with PNA at/above OP: ", "PNA at/above OP: "),
-        ("Reorder triggered after PNA fell below OP: ", "Below-OP reorder: "),
-        ("Guided reorders used: ", "Guided reorders: "),
-        ("Manual custom orders used: ", "Custom orders: "),
-        ("Parameter updates used: ", "Policy updates: "),
-        ("On-order inventory: ", "On order: "),
-        ("Avg inventory value: ", "Avg inv value: "),
-        ("Items at/below critical point: ", "At/below CP: "),
-        ("Items above surplus threshold: ", "Above surplus: "),
-        ("Backorder at close: ", "Close backorder: "),
-    )
-    for old, new in replacements:
-        if row.startswith(old):
-            return row.replace(old, new, 1)
-    return row
-
-
-def _lesson_item_snapshot_columns(level_index: int) -> tuple[str, ...]:
-    return {
-        1: ("item", "on_hand", "daily_usage", "backorder"),
-        2: ("item", "on_hand", "on_order", "backorder", "pna"),
-        3: ("item", "on_hand", "daily_usage", "backorder"),
-        4: ("item", "usage_rate", "hits_per_month", "on_hand"),
-        5: ("item", "usage_rate", "daily_usage", "on_hand"),
-        6: ("item", "on_hand", "on_order", "lead_time", "pna"),
-        7: ("item", "on_hand", "on_order", "backorder", "pna"),
-        8: ("item", "on_hand", "usage_rate", "op", "days_to_op"),
-        9: ("item", "safety_allowance", "op", "on_hand"),
-        10: ("item", "pna", "op", "safety_allowance", "soq"),
-        11: ("item", "hits_per_month", "usage_rate", "pna"),
-        12: ("item", "pna", "op", "lp", "soq"),
-        13: ("item", "pna", "op", "lp", "soq"),
-        14: ("item", "item_cost", "pna", "soq"),
-        15: ("item", "eoq", "oq", "pna", "op"),
-        16: ("item", "pna", "oq", "standard_pack", "soq"),
-        17: ("item", "pna", "cp", "surplus_line"),
-        18: ("item", "pna", "op", "lp", "soq", "on_order"),
-        19: ("item", "pna", "op", "lp", "soq"),
-    }.get(level_index, ("item", "on_hand", "on_order", "backorder"))
-
-
-def _lesson_item_snapshot_value(column: str, index: int, item: InventoryItem) -> str:
-    if column == "item":
-        return str(index)
-    if column == "on_hand":
-        return f"{item.on_hand:.1f}"
-    if column == "daily_usage":
-        return f"{item.daily_ur:.1f}"
-    if column == "usage_rate":
-        return f"{item.usage_rate:.1f}"
-    if column == "lead_time":
-        return f"{item.lead_time:.1f}"
-    if column == "hits_per_month":
-        return f"{item.hits_per_month:.1f}"
-    if column == "item_cost":
-        return format_money(item.item_cost)
-    if column == "on_order":
-        return f"{item_on_order(item):.1f}"
-    if column == "backorder":
-        return f"{item.backorder:.1f}"
-    if column == "pna":
-        return f"{item.pna:.1f}"
-    if column == "op":
-        return f"{item.op:.1f}"
-    if column == "lp":
-        return f"{item.lp:.1f}"
-    if column == "eoq":
-        return f"{item.eoq:.1f}"
-    if column == "oq":
-        return f"{item.oq:.1f}"
-    if column == "soq":
-        return f"{item.soq:.1f}"
-    if column == "standard_pack":
-        return f"{item.standard_pack:.1f}"
-    if column == "safety_allowance":
-        return f"{item.safety_allowance * 100.0:.1f}%"
-    if column == "cp":
-        return f"{item.cp:.1f}"
-    if column == "surplus_line":
-        return f"{item.surplus_line:.1f}"
-    if column == "days_to_op":
-        return f"{item.ats_days_frm_op:.1f}"
-    raise ValueError(f"Unsupported lesson snapshot column: {column}")
-
-
-def _lesson_item_snapshot_block(level_index: int, items: list[InventoryItem]) -> html.Div:
-    labels = {
-        "item": "Item",
-        "on_hand": "On Hand",
-        "daily_usage": "Daily Usage",
-        "usage_rate": "Usage",
-        "lead_time": "Lead Time",
-        "hits_per_month": "Hits",
-        "item_cost": "Cost",
-        "on_order": "On Order",
-        "backorder": "Backorder",
-        "pna": "PNA",
-        "op": "OP",
-        "lp": "LP",
-        "eoq": "EOQ",
-        "oq": "OQ",
-        "soq": "SOQ",
-        "standard_pack": "Pack",
-        "safety_allowance": "Safety %",
-        "cp": "CP",
-        "surplus_line": "Surplus threshold",
-        "days_to_op": "Days to OP",
-    }
-    columns = _lesson_item_snapshot_columns(level_index)
-    return _lesson_snapshot_table(
-        "Inventory",
-        tuple(labels[column] for column in columns),
-        tuple(
-            tuple(_lesson_item_snapshot_value(column, index, item) for column in columns)
-            for index, item in enumerate(items, start=1)
-        ),
-    )
-
-
-def _workspace_graph_height(state: SimulationState) -> int:
-    return {
-        "intro_pna": 340,
-        "workspace_signal": 300,
-        "workspace_advanced": 340,
-        "workspace_certification": 360,
-        "simulator": 460,
-    }.get(active_layout_variant(state), 400)
-
-
-def inventory_graph_height(state: SimulationState) -> int:
-    level = active_level(state)
-    if level is not None and level.index == 1:
-        return _LESSON_ONE_GRAPH_HEIGHT
-    if level is not None and level.index == 2:
-        return _LESSON_TWO_GRAPH_HEIGHT
-    if level is not None and level.index == 3:
-        return _FILL_RATE_GRAPH_HEIGHT
-    if level is not None and level.index in {12, 13}:
-        return 340
-    if level is not None and level.index == 17:
-        return 360
-    return _workspace_graph_height(state)
-
-
-def inventory_graph_style(state: SimulationState) -> dict[str, str]:
-    height = inventory_graph_height(state)
-    return {"height": f"{height}px", "minHeight": f"{height}px", "width": "100%"}
-
-
-def _workspace_grid_height(state: SimulationState, *, surface: str) -> str:
-    variant = active_layout_variant(state)
-    if surface == "inventory":
-        return {
-            "workspace_basic": "36rem",
-            "workspace_signal": "32rem",
-            "workspace_advanced": "28rem",
-            "workspace_certification": "26rem",
-            "simulator": "28rem",
-        }.get(variant, "28rem")
-    return {
-        "workspace_basic": "32rem",
-        "workspace_signal": "30rem",
-        "workspace_advanced": "26rem",
-        "workspace_certification": "26rem",
-        "simulator": "26rem",
-    }.get(variant, "26rem")
-
-
-def _lesson_uses_day_basis(level) -> bool:
-    day_basis_phrase = "day basis"
-    return day_basis_phrase in level.formula.casefold() or any(
-        day_basis_phrase in step.casefold() for step in level.tutorial_steps
-    )
-
-
-def _lesson_day_basis_helper(state: SimulationState) -> html.Div:
-    return html.Div(
-        f"Simulator simplified month: {state.global_settings.day_basis} days",
-        className="helper-copy mb-2",
-    )
-
-
-def _lesson_secondary_note(
-    title: str,
-    hint: str,
-    copy: str,
-) -> html.Details:
-    return _lesson_snapshot_disclosure(
-        html.Div(copy, className="helper-copy mb-0"),
-        label=title,
-        hint=hint,
-    )
-
-
-def _lesson_secondary_notes(level) -> list[html.Details]:
-    notes: list[html.Details] = []
-    if level.advanced_note:
-        notes.append(
-            _lesson_secondary_note(
-                "Advanced note",
-                "Optional deeper context",
-                level.advanced_note,
-            )
-        )
-    if level.csd_mapping_note:
-        notes.append(
-            _lesson_secondary_note(
-                "CSD mapping",
-                "Optional system mapping",
-                level.csd_mapping_note,
-            )
-        )
-    return notes
-
-
-def _projected_line_summary(
-    state: SimulationState,
-    *,
-    review_cycle: int,
-) -> tuple[float, int, float, float]:
-    settings = GlobalSettings.from_dict(asdict(state.global_settings))
-    settings.r_cycle = int(review_cycle)
-    settings.review_cycle_override_days = None
-    total_soq = 0.0
-    line_count = 0
-    total_lp = 0.0
-    total_cost = 0.0
-    for source in state.items:
-        item = InventoryItem.from_dict(source.to_dict())
-        update_gs_related_values(item, settings)
-        total_soq += item.soq
-        total_lp += item.lp
-        total_cost += item.soq * item.item_cost
-        if item.soq > 0:
-            line_count += 1
-    return total_soq, line_count, total_lp, total_cost
-
-
-def _emergency_replenishment_panel(state: SimulationState, level) -> html.Div:
-    normal_cycle = int(level.win_conditions.get("emergency_normal_review_cycle", 7))
-    emergency_cycle = int(level.win_conditions.get("emergency_review_cycle_min", 14))
-    normal_qty, normal_lines, normal_lp, normal_cost = _projected_line_summary(
-        state, review_cycle=normal_cycle
-    )
-    emergency_qty, emergency_lines, emergency_lp, emergency_cost = _projected_line_summary(
-        state, review_cycle=emergency_cycle
-    )
-    return html.Div(
-        [
-            html.Div(
-                (
-                    "A bridge buy is a temporary timing decision. Raising review cycle "
-                    "days makes the recommendation look farther ahead, which can bring "
-                    "near-line-point items into the same PO. Use it to cover the gap, "
-                    "then return to the regular cadence."
-                ),
-                className="helper-copy mb-2",
-            ),
-            _lesson_snapshot_table(
-                "Before / After RRAR",
-                ("Control", "Normal Buy", "Emergency Buy"),
-                (
-                    ("Review Cycle", f"{normal_cycle} days", f"{emergency_cycle} days"),
-                    ("Included Lines", str(normal_lines), str(emergency_lines)),
-                    ("Line Point Total", f"{normal_lp:.1f}", f"{emergency_lp:.1f}"),
-                    ("Suggested Qty", f"{normal_qty:.1f}", f"{emergency_qty:.1f}"),
-                    ("Extended Cost", format_money(normal_cost), format_money(emergency_cost)),
-                    ("Tradeoff", "Standard line coverage", "Higher bridge coverage"),
-                ),
-            ),
-            _lesson_snapshot_table(
-                "Demand Center Header",
-                ("Buyer", "Vendor", "Warehouse", "PLine", "Target", "Merge"),
-                (
-                    (
-                        "RSC",
-                        "100 - Apex Electrical",
-                        "MAIN",
-                        "ELEC",
-                        "$7,500",
-                        "Review after accept",
-                    ),
-                ),
-            ),
-        ],
-        className="lesson-snapshot-stack",
-    )
-
-
-def _signal_map_layout_signature(state: SimulationState, rows: pd.DataFrame) -> str:
-    effective_cycle = effective_review_cycle(state.global_settings)
-    return (
-        f"{active_layout_variant(state)}:items:{len(rows)}:"
-        f"r_cycle:{state.global_settings.r_cycle}:effective_cycle:{effective_cycle}"
-    )
-
-
-def _exception_map_layout_signature(rows: pd.DataFrame) -> str:
-    return f"exceptions:items:{len(rows)}"
-
-
-def _lesson_position_snapshot_block(state: SimulationState) -> html.Div:
-    variant = active_layout_variant(state)
-    on_order = sum(item_on_order(item) for item in state.items)
-    backorder = sum(item.backorder for item in state.items)
-    if variant == "workspace_basic":
-        at_or_below_op = sum(1 for item in state.items if item.on_hand <= item.op)
-        rows = (
-            (
-                f"{sum(item.on_hand for item in state.items):.1f}",
-                f"{on_order:.1f}",
-                f"{backorder:.1f}",
-                f"{at_or_below_op}/{len(state.items)}",
-            ),
-        )
-        return _lesson_snapshot_table(
-            "Position",
-            ("Available", "On Order", "Backorder", "At/Below OP"),
-            rows,
-        )
-    at_or_below_op = sum(1 for item in state.items if item.pna <= item.op)
-    rows = (
-        (
-            f"{sum(item.pna for item in state.items):.1f}",
-            f"{on_order:.1f}",
-            f"{backorder:.1f}",
-            f"{at_or_below_op}/{len(state.items)}",
-        ),
-    )
-    return _lesson_snapshot_table(
-        "Position",
-        ("PNA", "On Order", "Backorder", "At/Below OP"),
-        rows,
-    )
-
-
-_FIGURE_KIND_KEY = "figure_kind"
-_FIGURE_THEME_KEY = "theme"
-_FIGURE_LAYOUT_SIG_KEY = "layout_signature"
-
-
-def _grid_theme_class(theme: str) -> str:
-    return (
-        "ag-theme-quartz-dark imsim-ag-grid" if theme == "dark" else "ag-theme-quartz imsim-ag-grid"
-    )
-
-
-def po_overview_grid_options() -> dict[str, object]:
-    return {
-        "animateRows": False,
-        "rowSelection": {
-            "mode": "multiRow",
-            "checkboxes": True,
-            "headerCheckbox": True,
-            "enableClickSelection": True,
-            "enableSelectionWithoutKeys": True,
-        },
-        "selectionColumnDef": {
-            "width": 56,
-            "maxWidth": 56,
-            "resizable": False,
-            "sortable": False,
-            "pinned": "left",
-        },
-    }
-
-
-def custom_order_grid_options() -> dict[str, object]:
-    return {
-        "animateRows": False,
-        "singleClickEdit": True,
-        "stopEditingWhenCellsLoseFocus": True,
-        "enterNavigatesVerticallyAfterEdit": True,
-    }
-
-
-def _figure_meta(
-    kind: str,
-    theme: str,
-    *,
-    layout_signature: str = "static",
-) -> dict[str, str]:
-    return {
-        _FIGURE_KIND_KEY: kind,
-        _FIGURE_THEME_KEY: theme,
-        _FIGURE_LAYOUT_SIG_KEY: layout_signature,
-    }
-
-
-def _stable_history_points(state: SimulationState) -> list:
-    return state.history or []
-
-
-def _empty_marker_trace(
-    *,
-    name: str,
-    marker: dict[str, object],
-    hovertemplate: str,
-    customdata: list | None = None,
-) -> go.Scatter:
-    trace = go.Scatter(
-        x=[],
-        y=[],
-        mode="markers",
-        name=name,
-        marker=marker,
-        hovertemplate=hovertemplate,
-    )
-    if customdata is not None:
-        trace.customdata = customdata
-    return trace
-
-
-def _empty_line_trace(
-    *,
-    name: str,
-    line: dict[str, object],
-    marker: dict[str, object],
-    hovertemplate: str,
-) -> go.Scatter:
-    return go.Scatter(
-        x=[],
-        y=[],
-        mode="lines+markers",
-        name=name,
-        line=line,
-        marker=marker,
-        hovertemplate=hovertemplate,
-    )
-
-
-def _finalize_axes(
-    fig: go.Figure,
-    colors: dict[str, str],
-    *,
-    x_linear: bool = False,
-    x_range: list[float] | None = None,
-    y_tozero: bool = False,
-) -> go.Figure:
-    xaxis: dict[str, object] = {
-        "color": colors["text"],
-        "gridcolor": colors["line"],
-        "linecolor": colors["line"],
-        "title_font": {"color": colors["text"]},
-        "tickfont": {"color": colors["text"]},
-    }
-    if x_linear:
-        xaxis.update({"tickmode": "linear", "dtick": 1, "tick0": 1})
-    if x_range is not None:
-        xaxis["range"] = x_range
-        xaxis["tickformat"] = "d"
-    yaxis: dict[str, object] = {
-        "color": colors["text"],
-        "gridcolor": colors["line"],
-        "linecolor": colors["line"],
-        "title_font": {"color": colors["text"]},
-        "tickfont": {"color": colors["text"]},
-    }
-    if y_tozero:
-        yaxis["rangemode"] = "tozero"
-    else:
-        yaxis["zerolinecolor"] = colors["guide"]
-    fig.update_xaxes(**xaxis)
-    fig.update_yaxes(**yaxis)
-    return fig
-
-
-def _apply_signal_guides(
-    fig: go.Figure,
-    colors: dict[str, str],
-    *,
-    lower_label: str,
-    upper_label: str,
-    upper_y: float,
-) -> go.Figure:
-    fig.add_hline(
-        y=0,
-        line_dash="dot",
-        line_color=colors["guide"],
-        annotation_text=lower_label,
-        annotation_font_color=colors["text"],
-    )
-    fig.add_hline(
-        y=upper_y,
-        line_dash="dot",
-        line_color=colors["guide"],
-        annotation_text=upper_label,
-        annotation_font_color=colors["text"],
-    )
-    return fig
-
-
-def _apply_single_guide(
-    fig: go.Figure,
-    colors: dict[str, str],
-    *,
-    y: float,
-    label: str,
-) -> go.Figure:
-    fig.add_hline(
-        y=y,
-        line_dash="dot",
-        line_color=colors["guide"],
-        annotation_text=label,
-        annotation_font_color=colors["text"],
-    )
-    return fig
-
-
-def _lesson_one_figure(state: SimulationState, theme: str, colors: dict[str, str]) -> go.Figure:
-    history = _stable_history_points(state)
-    days = [point.day for point in history]
-    on_hand = [point.total_on_hand for point in history]
-    backorder = [point.total_backorder for point in history]
-    fig = go.Figure(
-        data=[
-            _empty_line_trace(
-                name="On Hand",
-                line=_plot_line(colors["pna"]),
-                marker=_plot_marker(colors["pna"], _PLOT_MARKER_SIZE_REM["lesson"]),
-                hovertemplate="On hand %{y:.1f}<extra></extra>",
-            ),
-            _empty_line_trace(
-                name="Backorder",
-                line=_plot_line(colors["zero"], dash="dash"),
-                marker=_plot_marker(colors["zero"], _PLOT_MARKER_SIZE_REM["lesson_backorder"]),
-                hovertemplate="Backorder %{y:.1f}<extra></extra>",
-            ),
-        ]
-    )
-    fig.update_layout(
-        **_plot_base_layout(None, colors, height=_LESSON_ONE_GRAPH_HEIGHT),
-        xaxis_title="Day",
-        yaxis_title="Units",
-        hovermode="x unified",
-        uirevision=f"lesson-1:{theme}",
-        meta=_figure_meta("lesson-1", theme),
-    )
-    fig.update_traces(x=days, selector={"name": "On Hand"})
-    fig.update_traces(y=on_hand, selector={"name": "On Hand"})
-    fig.update_traces(x=days, selector={"name": "Backorder"})
-    fig.update_traces(y=backorder, selector={"name": "Backorder"})
-    _apply_single_guide(fig, colors, y=0, label="Zero on hand")
-    return _finalize_axes(fig, colors, x_linear=True, y_tozero=True)
-
-
-def _lesson_two_figure(state: SimulationState, theme: str, colors: dict[str, str]) -> go.Figure:
-    history = _stable_history_points(state)
-    days = [point.day for point in history]
-    op_value = state.items[0].op if state.items else 0.0
-    fig = go.Figure(
-        data=[
-            _empty_line_trace(
-                name="On Hand",
-                line=_plot_line(colors["pna"]),
-                marker=_plot_marker(colors["pna"], _PLOT_MARKER_SIZE_REM["lesson"]),
-                hovertemplate="On hand %{y:.1f}<extra></extra>",
-            ),
-            _empty_line_trace(
-                name="On Order",
-                line=_plot_line(colors["ats"]),
-                marker=_plot_marker(colors["ats"], _PLOT_MARKER_SIZE_REM["lesson"]),
-                hovertemplate="On order %{y:.1f}<extra></extra>",
-            ),
-            _empty_line_trace(
-                name="PNA",
-                line=_plot_line(colors["proposed"]),
-                marker=_plot_marker(colors["proposed"], _PLOT_MARKER_SIZE_REM["lesson"]),
-                hovertemplate="PNA %{y:.1f}<extra></extra>",
-            ),
-            _empty_line_trace(
-                name="OP",
-                line=_plot_line(colors["guide"], dash="dot"),
-                marker=_plot_marker(colors["guide"], _PLOT_MARKER_SIZE_REM["lesson_backorder"]),
-                hovertemplate="OP %{y:.1f}<extra></extra>",
-            ),
-            _empty_line_trace(
-                name="Backorder",
-                line=_plot_line(colors["zero"], dash="dash"),
-                marker=_plot_marker(colors["zero"], _PLOT_MARKER_SIZE_REM["lesson_backorder"]),
-                hovertemplate="Backorder %{y:.1f}<extra></extra>",
-            ),
-        ]
-    )
-    fig.update_layout(
-        **_plot_base_layout(None, colors, height=_LESSON_TWO_GRAPH_HEIGHT),
-        xaxis_title="Day",
-        yaxis_title="Units",
-        hovermode="x unified",
-        uirevision=f"lesson-2:{theme}",
-        meta=_figure_meta("lesson-2", theme),
-    )
-    fig.update_traces(x=days, selector={"name": "On Hand"})
-    fig.update_traces(y=[point.total_on_hand for point in history], selector={"name": "On Hand"})
-    fig.update_traces(x=days, selector={"name": "On Order"})
-    fig.update_traces(
-        y=[point.total_on_order for point in history],
-        selector={"name": "On Order"},
-    )
-    fig.update_traces(x=days, selector={"name": "PNA"})
-    fig.update_traces(y=[point.total_pna for point in history], selector={"name": "PNA"})
-    fig.update_traces(x=days, selector={"name": "OP"})
-    fig.update_traces(y=[op_value for _day in days], selector={"name": "OP"})
-    fig.update_traces(x=days, selector={"name": "Backorder"})
-    fig.update_traces(
-        y=[point.total_backorder for point in history],
-        selector={"name": "Backorder"},
-    )
-    _apply_single_guide(fig, colors, y=0, label="Zero")
-    return _finalize_axes(fig, colors, x_linear=True, y_tozero=True)
-
-
-def _fill_rate_figure(state: SimulationState, theme: str, colors: dict[str, str]) -> go.Figure:
-    level = active_level(state)
-    target = float((level.win_conditions if level is not None else {}).get("fill_rate_min", 0.0))
-    current = fill_rate(state)
-    fill_pct = 0.0 if current is None else current * 100.0
-    target_pct = target * 100.0
-    complete = max(0, state.service_totals.orders - state.service_totals.orders_stockout)
-    incomplete = state.service_totals.orders_stockout
-    fig = go.Figure(
-        data=[
-            go.Indicator(
-                mode="gauge+number",
-                value=fill_pct,
-                number={"suffix": "%", "font": {"color": colors["text"]}},
-                title={"text": "Cumulative Fill Rate", "font": {"color": colors["text"]}},
-                gauge={
-                    "axis": {"range": [0, 100], "tickcolor": colors["text"]},
-                    "bar": {"color": colors["pna"]},
-                    "bgcolor": colors["plot_bg"],
-                    "bordercolor": colors["line"],
-                    "steps": [
-                        {"range": [0, target_pct], "color": "rgba(248, 113, 113, 0.18)"},
-                        {"range": [target_pct, 100], "color": "rgba(45, 212, 191, 0.18)"},
-                    ],
-                    "threshold": {
-                        "line": {"color": colors["zero"], "width": 4},
-                        "thickness": 0.8,
-                        "value": target_pct,
-                    },
-                },
-                domain={"x": [0, 0.58], "y": [0, 1]},
-            ),
-            go.Bar(
-                x=["Complete", "Incomplete"],
-                y=[complete, incomplete],
-                marker_color=[colors["pna"], colors["zero"]],
-                text=[complete, incomplete],
-                textposition="auto",
-                hovertemplate="%{x} lines: %{y}<extra></extra>",
-                xaxis="x2",
-                yaxis="y2",
-                showlegend=False,
-            ),
-        ]
-    )
-    base_layout = _plot_base_layout(None, colors, height=_FILL_RATE_GRAPH_HEIGHT)
-    base_layout["margin"] = {
-        "l": _rem_to_px(1.5),
-        "r": _rem_to_px(1.5),
-        "t": _rem_to_px(2.6),
-        "b": _rem_to_px(2.0),
-    }
-    fig.update_layout(
-        **base_layout,
-        uirevision=f"lesson-3:{theme}",
-        meta=_figure_meta("lesson-3-fill-rate", theme),
-        xaxis2={
-            "domain": [0.7, 1.0],
-            "anchor": "y2",
-            "color": colors["text"],
-            "gridcolor": colors["line"],
-        },
-        yaxis2={
-            "domain": [0.24, 0.84],
-            "anchor": "x2",
-            "color": colors["text"],
-            "gridcolor": colors["line"],
-            "rangemode": "tozero",
-            "title": "Lines",
-        },
-    )
-    fig.update_traces(domain={"x": [0.0, 0.6], "y": [0.08, 0.94]}, selector={"type": "indicator"})
-    return fig
-
-
-def _signal_map_figure(state: SimulationState, theme: str, colors: dict[str, str]) -> go.Figure:
-    rows = _items_frame(state.items)
-    if rows.empty:
-        rows = pd.DataFrame(
-            columns=[
-                "idx",
-                "pna_days_frm_op",
-                "pro_pna_days_frm_op",
-                "pna",
-                "lp",
-                "ats_days_frm_op",
-                "ats_days_to_stockout",
-                "no_pna_days_frm_op",
-                "stockout_today",
-            ]
-        )
-    item_numbers = (rows["idx"] + 1).tolist() if "idx" in rows else []
-    hover = "Item %{x}<br>%{y:.1f} days<extra>%{fullData.name}</extra>"
-    proposed_mask = (
-        (rows["pro_pna_days_frm_op"] != rows["pna_days_frm_op"]) & (rows["pna"] <= rows["lp"])
-        if not rows.empty
-        else pd.Series(dtype=bool)
-    )
-    proposed_rows = rows[proposed_mask] if not rows.empty else rows
-    stockout_rows = rows[rows["stockout_today"]] if not rows.empty else rows
-    fig = go.Figure(
-        data=[
-            _empty_marker_trace(
-                name="PNA",
-                marker=_plot_marker(colors["pna"], _PLOT_MARKER_SIZE_REM["signal"]),
-                hovertemplate=hover,
-            ),
-            _empty_marker_trace(
-                name="PNA + SOQ",
-                marker=_plot_marker(
-                    colors["proposed"],
-                    _PLOT_MARKER_SIZE_REM["signal_proposed"],
-                    symbol="circle-open",
-                    line=_plot_marker_outline(colors["proposed"]),
-                ),
-                hovertemplate=hover,
-            ),
-            _empty_marker_trace(
-                name="Available to Sell",
-                marker=_plot_marker(
-                    colors["ats"],
-                    _PLOT_MARKER_SIZE_REM["signal_ats"],
-                    symbol="x",
-                ),
-                hovertemplate=(
-                    "Item %{x}<br>%{customdata:.1f} days to stockout"
-                    "<extra>Available to Sell</extra>"
-                ),
-                customdata=[],
-            ),
-            _empty_marker_trace(
-                name="0 PNA",
-                marker=_plot_marker(colors["zero"], _PLOT_MARKER_SIZE_REM["signal_zero"]),
-                hovertemplate=hover,
-            ),
-            _empty_marker_trace(
-                name="Stockout Today",
-                marker=_plot_marker(
-                    colors["zero"],
-                    _PLOT_MARKER_SIZE_REM["signal_stockout"],
-                    symbol="circle-open-dot",
-                    line=_plot_marker_outline(colors["zero"]),
-                ),
-                hovertemplate=hover,
-            ),
-        ]
-    )
-    fig.update_layout(
-        **_plot_base_layout(
-            None,
-            colors,
-            include_margin=not rows.empty,
-            height=inventory_graph_height(state),
-        ),
-        xaxis_title="Item",
-        yaxis_title="Days from OP",
-        hovermode="closest",
-        uirevision=f"signal-map:{theme}",
-        meta=_figure_meta(
-            "signal-map",
-            theme,
-            layout_signature=_signal_map_layout_signature(state, rows),
-        ),
-    )
-    fig.update_traces(x=item_numbers, selector={"name": "PNA"})
-    fig.update_traces(y=rows["pna_days_frm_op"].tolist(), selector={"name": "PNA"})
-    fig.update_traces(
-        x=(proposed_rows["idx"] + 1).tolist() if not proposed_rows.empty else [],
-        selector={"name": "PNA + SOQ"},
-    )
-    fig.update_traces(
-        y=proposed_rows["pro_pna_days_frm_op"].tolist() if not proposed_rows.empty else [],
-        selector={"name": "PNA + SOQ"},
-    )
-    fig.update_traces(x=item_numbers, selector={"name": "Available to Sell"})
-    fig.update_traces(y=rows["ats_days_frm_op"].tolist(), selector={"name": "Available to Sell"})
-    fig.update_traces(
-        customdata=rows["ats_days_to_stockout"].tolist(),
-        selector={"name": "Available to Sell"},
-    )
-    fig.update_traces(x=item_numbers, selector={"name": "0 PNA"})
-    fig.update_traces(y=rows["no_pna_days_frm_op"].tolist(), selector={"name": "0 PNA"})
-    fig.update_traces(
-        x=(stockout_rows["idx"] + 1).tolist() if not stockout_rows.empty else [],
-        selector={"name": "Stockout Today"},
-    )
-    fig.update_traces(
-        y=stockout_rows["pna_days_frm_op"].tolist() if not stockout_rows.empty else [],
-        selector={"name": "Stockout Today"},
-    )
-    _apply_signal_guides(
-        fig,
-        colors,
-        lower_label="OP",
-        upper_label=(
-            f"LP Override ({effective_review_cycle(state.global_settings)}d)"
-            if state.global_settings.review_cycle_override_days is not None
-            else "LP"
-        ),
-        upper_y=effective_review_cycle(state.global_settings),
-    )
-    return _finalize_axes(
-        fig,
-        colors,
-        x_linear=True,
-        x_range=[0.5, max(1, len(rows)) + 0.5],
-    )
-
-
-def _exception_signal_figure(
-    state: SimulationState,
-    theme: str,
-    colors: dict[str, str],
-) -> go.Figure:
-    rows = _items_frame(state.items)
-    if rows.empty:
-        rows = pd.DataFrame(columns=["idx", "pna", "cp", "surplus_line", "on_hand"])
-    item_numbers = (rows["idx"] + 1).tolist() if "idx" in rows else []
-    hover = "Item %{x}<br>%{y:.1f} units<extra>%{fullData.name}</extra>"
-    fig = go.Figure(
-        data=[
-            _empty_marker_trace(
-                name="PNA",
-                marker=_plot_marker(colors["pna"], _PLOT_MARKER_SIZE_REM["signal"]),
-                hovertemplate=hover,
-            ),
-            _empty_marker_trace(
-                name="Critical Point",
-                marker=_plot_marker(
-                    colors["zero"],
-                    _PLOT_MARKER_SIZE_REM["signal_zero"],
-                    symbol="diamond-open",
-                    line=_plot_marker_outline(colors["zero"]),
-                ),
-                hovertemplate=hover,
-            ),
-            _empty_marker_trace(
-                name="Surplus Threshold",
-                marker=_plot_marker(
-                    colors["proposed"],
-                    _PLOT_MARKER_SIZE_REM["signal_proposed"],
-                    symbol="circle-open",
-                    line=_plot_marker_outline(colors["proposed"]),
-                ),
-                hovertemplate=hover,
-            ),
-            _empty_marker_trace(
-                name="On Hand",
-                marker=_plot_marker(
-                    colors["ats"],
-                    _PLOT_MARKER_SIZE_REM["signal_ats"],
-                    symbol="x",
-                ),
-                hovertemplate=hover,
-            ),
-        ]
-    )
-    fig.update_layout(
-        **_plot_base_layout(
-            None,
-            colors,
-            include_margin=not rows.empty,
-            height=inventory_graph_height(state),
-        ),
-        xaxis_title="Item",
-        yaxis_title="Units",
-        hovermode="closest",
-        uirevision=f"exception-map:{theme}",
-        meta=_figure_meta(
-            "exception-map",
-            theme,
-            layout_signature=_exception_map_layout_signature(rows),
-        ),
-    )
-    fig.update_traces(x=item_numbers, selector={"name": "PNA"})
-    fig.update_traces(y=rows["pna"].tolist(), selector={"name": "PNA"})
-    fig.update_traces(x=item_numbers, selector={"name": "Critical Point"})
-    fig.update_traces(y=rows["cp"].tolist(), selector={"name": "Critical Point"})
-    fig.update_traces(x=item_numbers, selector={"name": "Surplus Threshold"})
-    fig.update_traces(y=rows["surplus_line"].tolist(), selector={"name": "Surplus Threshold"})
-    fig.update_traces(x=item_numbers, selector={"name": "On Hand"})
-    fig.update_traces(y=rows["on_hand"].tolist(), selector={"name": "On Hand"})
-    _apply_single_guide(fig, colors, y=0, label="Zero")
-    return _finalize_axes(
-        fig,
-        colors,
-        x_linear=True,
-        x_range=[0.5, max(1, len(rows)) + 0.5],
-        y_tozero=True,
-    )
-
-
-def build_inventory_figure(state: SimulationState, theme: str = "light") -> go.Figure:
-    colors = _figure_theme(theme)
-    level = active_level(state)
-    if level is not None and level.index == 1:
-        return _lesson_one_figure(state, theme, colors)
-    if level is not None and level.index == 2:
-        return _lesson_two_figure(state, theme, colors)
-    if level is not None and level.index == 3:
-        return _fill_rate_figure(state, theme, colors)
-    if level is not None and level.index == 17:
-        return _exception_signal_figure(state, theme, colors)
-    return _signal_map_figure(state, theme, colors)
-
-
-def refresh_inventory_figure(
-    state: SimulationState,
-    theme: str = "light",
-    current_figure: dict | None = None,
-) -> go.Figure | Patch:
-    target = build_inventory_figure(state, theme)
-    layout = (current_figure or {}).get("layout") or {}
-    current_meta = layout.get("meta") or {}
-    target_meta = dict(target.layout.meta or {})
-    if (
-        current_meta.get(_FIGURE_KIND_KEY) != target_meta.get(_FIGURE_KIND_KEY)
-        or current_meta.get(_FIGURE_THEME_KEY) != target_meta.get(_FIGURE_THEME_KEY)
-        or current_meta.get(_FIGURE_LAYOUT_SIG_KEY) != target_meta.get(_FIGURE_LAYOUT_SIG_KEY)
-        or len((current_figure or {}).get("data") or []) != len(target.data)
-    ):
-        return target
-
-    patched = Patch()
-    for index, trace in enumerate(target.data):
-        if getattr(trace, "x", None) is not None:
-            patched["data"][index]["x"] = list(trace.x)
-        if getattr(trace, "y", None) is not None:
-            patched["data"][index]["y"] = list(trace.y)
-        if getattr(trace, "value", None) is not None:
-            patched["data"][index]["value"] = trace.value
-        if getattr(trace, "text", None) is not None:
-            patched["data"][index]["text"] = (
-                trace.text if isinstance(trace.text, str) else list(trace.text)
-            )
-        if getattr(trace, "customdata", None) is not None:
-            patched["data"][index]["customdata"] = list(trace.customdata)
-        if getattr(trace, "hovertemplate", None) is not None:
-            patched["data"][index]["hovertemplate"] = trace.hovertemplate
-    return patched
-
-
-def service_card_children(state: SimulationState) -> list:
-    level = active_level(state)
-    today = state.service_today
-    totals = state.service_totals
-    ats = sum(item.on_hand for item in state.items)
-    on_order = sum(item_on_order(item) for item in state.items)
-    backorder = sum(item.backorder for item in state.items)
-    fill_today = (
-        None if today.orders == 0 else 100.0 * (today.orders - today.orders_stockout) / today.orders
-    )
-    fill_total = (
-        None
-        if totals.orders == 0
-        else 100.0 * (totals.orders - totals.orders_stockout) / totals.orders
-    )
-    if level is not None and level.index == 1 and state.items:
-        item = state.items[0]
-        status = "Depleted" if item.on_hand <= 0 else "Available"
-        return [
-            html.Div(
-                [
-                    _lesson_snapshot_table(
-                        "Service",
-                        ("On Hand", "Daily Usage", "Backorder", "Status"),
-                        (
-                            (
-                                f"{item.on_hand:.1f} units",
-                                f"{item.daily_ur:.1f} units/day",
-                                f"{item.backorder:.1f} units",
-                                status,
-                            ),
-                        ),
-                    ),
-                    _lesson_snapshot_table(
-                        "Inventory",
-                        ("Item", "On Hand", "Daily Usage", "Backorder"),
-                        (
-                            (
-                                "1",
-                                f"{item.on_hand:.0f}",
-                                f"{item.daily_ur:.0f}",
-                                f"{item.backorder:.0f}",
-                            ),
-                        ),
-                    ),
-                ],
-                className="lesson-snapshot-stack",
-            )
-        ]
-    if level is not None and level.index == 2 and state.items:
-        item = state.items[0]
-        pna_at_or_above_op = sum(1 for candidate in state.items if candidate.pna >= candidate.op)
-        reorder_target = int(level.win_conditions.get("guided_order_below_op_min", 1))
-        snapshot_body = html.Div(
-            [
-                _lesson_snapshot_table(
-                    "PNA Position",
-                    ("On Hand", "On Order", "Backorder", "PNA", "OP"),
-                    (
-                        (
-                            f"{item.on_hand:.1f}",
-                            f"{item_on_order(item):.1f}",
-                            f"{item.backorder:.1f}",
-                            f"{item.pna:.1f}",
-                            f"{item.op:.1f}",
-                        ),
-                    ),
-                ),
-                _lesson_snapshot_table(
-                    "Goal Check",
-                    ("Close PNA >= OP", "Below-OP Reorders", "Days Left"),
-                    (
-                        (
-                            f"{pna_at_or_above_op}/{len(state.items)}",
-                            f"{state.training.guided_orders_below_op}/{reorder_target}",
-                            str(lesson_days_remaining(state)),
-                        ),
-                    ),
-                ),
-            ],
-            className="lesson-snapshot-grid",
-        )
-        return [
-            _lesson_snapshot_disclosure(
-                snapshot_body,
-                hint="PNA and goal status",
-            )
-        ]
-    if level is not None and level.index == 3 and state.items:
-        snapshot_body = html.Div(
-            [
-                html.Div(
-                    [
-                        _lesson_snapshot_table(
-                            "Service",
-                            ("Orders", "Stockouts", "Today Fill", "Total Fill", "Zero Available"),
-                            (
-                                (
-                                    str(today.orders),
-                                    str(today.orders_stockout),
-                                    _format_pct(fill_today),
-                                    _format_pct(fill_total),
-                                    str(today.zero_on_hand_hits),
-                                ),
-                            ),
-                        ),
-                        _lesson_position_snapshot_block(state),
-                    ],
-                    className="lesson-snapshot-grid",
-                ),
-                _lesson_item_snapshot_block(level.index, state.items),
-            ],
-            className="lesson-snapshot-stack",
-        )
-        return [
-            _lesson_snapshot_disclosure(
-                snapshot_body,
-                hint="Service and order-point status",
-            )
-        ]
-    if level is not None and state.items:
-        top_blocks = [
-            _lesson_snapshot_table(
-                "Service",
-                ("Orders", "Stockouts", "Today Fill", "Total Fill", "Zero Available"),
-                (
-                    (
-                        str(today.orders),
-                        str(today.orders_stockout),
-                        _format_pct(fill_today),
-                        _format_pct(fill_total),
-                        str(today.zero_on_hand_hits),
-                    ),
-                ),
-            ),
-            _lesson_position_snapshot_block(state),
-        ]
-        return [
-            _lesson_snapshot_disclosure(
-                html.Div(
-                    [
-                        html.Div(top_blocks, className="lesson-snapshot-grid"),
-                        _lesson_item_snapshot_block(level.index, state.items),
-                    ],
-                    className="lesson-snapshot-stack",
-                ),
-                hint="Service, position, and item status",
-            )
-        ]
-
-    return [
-        dbc.ListGroup(
-            [
-                dbc.ListGroupItem(
-                    f"Today: Orders {today.orders} • Stockouts {today.orders_stockout} • "
-                    f"Zero available hits "
-                    f"{today.zero_on_hand_hits}"
-                ),
-                dbc.ListGroupItem(
-                    f"Fill Rate (cumulative): {_format_pct(fill_total)}  •  "
-                    f"Today: {_format_pct(fill_today)}"
-                ),
-                dbc.ListGroupItem(
-                    dbc.Badge(
-                        f"Available {int(ats)} • On-order {int(on_order)} • Backorder "
-                        f"{int(backorder)}",
-                        color="secondary",
-                        pill=True,
-                    )
-                ),
-            ],
-            flush=True,
-        )
-    ]
-
-
-def costs_card_children(state: SimulationState) -> list:
-    costs = state.costs
-    total = costs.ordering + costs.holding + costs.stockout + costs.expedite
-    return [
-        dbc.ListGroup(
-            [
-                dbc.ListGroupItem(f"Inventory Overhead: {format_money(total)}"),
-                dbc.ListGroupItem(
-                    " • ".join(
-                        [
-                            f"Ordering {format_money(costs.ordering)}",
-                            f"Holding {format_money(costs.holding)}",
-                            f"Stockout {format_money(costs.stockout)}",
-                            f"Expedite {format_money(costs.expedite)}",
-                        ]
-                    )
-                ),
-                dbc.ListGroupItem(f"Receipts Booked: {format_money(costs.purchases)}"),
-            ],
-            flush=True,
-        )
-    ]
-
-
-def sales_card_children(state: SimulationState) -> list:
-    sales = state.sales
-    costs = state.costs
-    gross = sales.revenue - sales.cogs
-    gross_pct = None if sales.revenue <= 0 else 100.0 * gross / sales.revenue
-    after_overhead = gross - costs.total
-    after_pct = None if sales.revenue <= 0 else 100.0 * after_overhead / sales.revenue
-
-    def fmt(value: float | None) -> str:
-        return "—" if value is None else f"{value:.1f}%"
-
-    return [
-        dbc.ListGroup(
-            [
-                dbc.ListGroupItem(f"Sales: {format_money(sales.revenue)}"),
-                dbc.ListGroupItem(
-                    f"COGS {format_money(sales.cogs)} • "
-                    f"GM$ {format_money(gross)} • GM% {fmt(gross_pct)}"
-                ),
-                dbc.ListGroupItem(
-                    f"After overhead {format_money(after_overhead)} • {fmt(after_pct)}"
-                ),
-            ],
-            flush=True,
-        )
-    ]
-
-
-def _kpi_card(title: str, value: str, tone: str, subtitle: str) -> dbc.Card:
-    return dbc.Card(
-        dbc.CardBody(
-            [
-                html.Div(title, className="metric-label"),
-                html.Div(value, className="metric-value"),
-                html.Div(subtitle, className="metric-subtitle"),
-            ]
-        ),
-        className=f"metric-card metric-{tone}",
-    )
-
-
-def build_kpi_strip(state: SimulationState) -> list:
-    revenue = state.sales.revenue
-    gross = revenue - state.sales.cogs
-    avg_inventory = state.analytics.inv_value_daysum / max(1, state.day)
-    fill_rate = (
-        None
-        if state.service_totals.orders == 0
-        else (state.service_totals.orders - state.service_totals.orders_stockout)
-        / state.service_totals.orders
-    )
-    turns = None if avg_inventory <= 0 else state.sales.cogs / avg_inventory
-    gmroi = None if avg_inventory <= 0 else gross / avg_inventory
-    after_overhead = gross - state.costs.total
-    after_pct = None if revenue <= 0 else after_overhead / revenue
-
-    def fmt_pct(value: float | None) -> str:
-        return "—" if value is None else f"{value * 100:.1f}%"
-
-    def fmt_ratio(value: float | None) -> str:
-        return "—" if value is None else f"{value:.2f}"
-
-    return [
-        dbc.Row(
-            [
-                dbc.Col(
-                    _kpi_card("Day", f"{state.day}", "neutral", "Simulation clock"),
-                    md=6,
-                    xl=2,
-                ),
-                dbc.Col(
-                    _kpi_card(
-                        "Fill Rate",
-                        fmt_pct(fill_rate),
-                        "good" if (fill_rate or 0) >= 0.95 else "warn",
-                        "Cumulative service",
-                    ),
-                    md=6,
-                    xl=2,
-                ),
-                dbc.Col(
-                    _kpi_card(
-                        "After-Overhead GM",
-                        fmt_pct(after_pct),
-                        "good" if (after_pct or 0) >= 0.15 else "warn",
-                        format_money(after_overhead),
-                    ),
-                    md=6,
-                    xl=2,
-                ),
-                dbc.Col(
-                    _kpi_card(
-                        "Avg Inventory",
-                        format_money(avg_inventory),
-                        "neutral",
-                        "Daily midpoint cost",
-                    ),
-                    md=6,
-                    xl=2,
-                ),
-                dbc.Col(
-                    _kpi_card("Turns", fmt_ratio(turns), "neutral", "COGS / avg inventory"),
-                    md=6,
-                    xl=2,
-                ),
-                dbc.Col(
-                    _kpi_card("GMROI", fmt_ratio(gmroi), "neutral", "Gross margin return"),
-                    md=6,
-                    xl=2,
-                ),
-            ],
-            className="g-3",
-        )
-    ]
-
-
-def build_po_overview_rows(state: SimulationState) -> list[dict[str, int | float | str]]:
-    rows: list[dict[str, int | float | str]] = []
-    for item_index, item in enumerate(state.items):
-        for receipt in item.pipeline:
-            days_left = max(0, receipt.eta_day - state.day)
-            rows.append(
-                {
-                    "item": item_index + 1,
-                    "receipt_id": receipt.receipt_id,
-                    "qty": int(receipt.qty),
-                    "eta_day": int(receipt.eta_day),
-                    "days_left": days_left,
-                }
-            )
-    return rows
-
-
-def build_po_overview_grid(state: SimulationState, theme: str = "light") -> dag.AgGrid | dbc.Alert:
-    rows = build_po_overview_rows(state)
-    if not rows:
-        return dbc.Alert("No open purchase orders.", color="secondary")
-    return dag.AgGrid(
-        id="po-overview-grid",
-        rowData=rows,
-        columnDefs=[
-            {"field": "item", "headerName": "Item", "pinned": "left", "maxWidth": 90},
-            {"field": "receipt_id", "headerName": "PO Line"},
-            {"field": "qty", "headerName": "Qty", "type": "numericColumn"},
-            {"field": "eta_day", "headerName": "ETA", "type": "numericColumn"},
-            {"field": "days_left", "headerName": "Days Left", "type": "numericColumn"},
-        ],
-        defaultColDef={"sortable": True, "filter": True, "resizable": True},
-        className=_grid_theme_class(theme),
-        columnSize="sizeToFit",
-        dashGridOptions=po_overview_grid_options(),
-        style={"height": _workspace_grid_height(state, surface="modal"), "width": "100%"},
-    )
-
-
-def build_custom_order_rows(state: SimulationState) -> list[dict[str, int | float]]:
-    rows: list[dict[str, int | float]] = []
-    for index, item in enumerate(state.items, start=1):
-        rows.append(
-            {
-                "item_index": index - 1,
-                "item": index,
-                "on_hand": int(item.on_hand),
-                "on_order": int(item_on_order(item)),
-                "backorder": int(item.backorder),
-                "usage_rate": round(item.usage_rate),
-                "lead_time": round(item.lead_time),
-                "op": round(item.op),
-                "lp": round(item.lp),
-                "oq": round(item.oq),
-                "order_qty": int(round(item.soq)),
-            }
-        )
-    return rows
-
-
-def build_custom_order_grid(state: SimulationState, theme: str = "light") -> dag.AgGrid | dbc.Alert:
-    rows = build_custom_order_rows(state)
-    if not rows:
-        return dbc.Alert("No items available.", color="warning")
-    return dag.AgGrid(
-        id="custom-order-grid",
-        rowData=rows,
-        columnDefs=[
-            {"field": "item", "headerName": "Item", "pinned": "left", "maxWidth": 90},
-            {"field": "on_hand", "headerName": "ATS (On Hand)", "type": "numericColumn"},
-            {"field": "on_order", "headerName": "On-Order", "type": "numericColumn"},
-            {"field": "backorder", "headerName": "Backorder", "type": "numericColumn"},
-            {"field": "usage_rate", "headerName": "Usage", "type": "numericColumn"},
-            {"field": "lead_time", "headerName": "Lead Time", "type": "numericColumn"},
-            {"field": "op", "headerName": "OP", "type": "numericColumn"},
-            {"field": "lp", "headerName": "LP", "type": "numericColumn"},
-            {"field": "oq", "headerName": "OQ", "type": "numericColumn"},
-            {
-                "field": "order_qty",
-                "headerName": "Order Qty",
-                "type": "numericColumn",
-                "editable": True,
-                "cellClass": "custom-order-qty-cell",
-                "cellEditor": "agNumberCellEditor",
-                "cellEditorParams": {
-                    "min": 0,
-                    "step": 1,
-                    "showStepperButtons": True,
-                },
-                "headerTooltip": (
-                    "Current suggested order quantity (SOQ); adjust before placing "
-                    "the custom order."
-                ),
-            },
-        ],
-        defaultColDef={"sortable": True, "filter": True, "resizable": True},
-        className=_grid_theme_class(theme),
-        columnSize="sizeToFit",
-        dashGridOptions=custom_order_grid_options(),
-        style={"height": _workspace_grid_height(state, surface="modal"), "width": "100%"},
-    )
-
-
-def build_inventory_table(state: SimulationState, theme: str = "light"):
-    level = active_level(state)
-    column_config = {
-        "item": ("item", "Item"),
-        "usage_rate": ("usage_rate", "Usage"),
-        "hits_per_month": ("hits_per_month", "Hits"),
-        "item_cost": ("item_cost", "Cost"),
-        "lead_time": ("lead_time", "Lead Time"),
-        "op": ("op", "OP"),
-        "lp": ("lp", "LP"),
-        "eoq": ("eoq", "EOQ"),
-        "oq": ("oq", "OQ"),
-        "pna": ("pna", "PNA"),
-        "on_hand": ("on_hand", "On Hand"),
-        "on_order": ("on_order", "On Order"),
-        "backorder": ("backorder", "Backorder"),
-        "soq": ("soq", "SOQ"),
-        "standard_pack": ("standard_pack", "Pack"),
-        "safety_allowance": ("safety_allowance", "Safety %"),
-        "cp": ("cp", "CP"),
-        "surplus_line": ("surplus_line", "Surplus threshold"),
-        "days_to_op": ("days_to_op", "Days to OP"),
-        "daily_usage": ("daily_usage", "Daily Usage"),
-    }
-    rows = []
-    for index, item in enumerate(state.items, start=1):
-        rows.append(
-            {
-                "item": index,
-                "usage_rate": round(item.usage_rate, 2),
-                "hits_per_month": round(item.hits_per_month, 2),
-                "item_cost": round(item.item_cost, 2),
-                "lead_time": round(item.lead_time, 2),
-                "op": round(item.op, 2),
-                "lp": round(item.lp, 2),
-                "eoq": round(item.eoq, 2),
-                "oq": round(item.oq, 2),
-                "pna": round(item.pna, 2),
-                "on_hand": round(item.on_hand, 2),
-                "on_order": round(item_on_order(item), 2),
-                "backorder": round(item.backorder, 2),
-                "soq": round(item.soq, 2),
-                "standard_pack": round(item.standard_pack, 2),
-                "safety_allowance": round(item.safety_allowance * 100.0, 1),
-                "cp": round(item.cp, 2),
-                "surplus_line": round(item.surplus_line, 2),
-                "days_to_op": round(item.ats_days_frm_op, 2),
-                "daily_usage": round(item.daily_ur, 2),
-            }
-        )
-    if not rows:
-        return dbc.Alert(
-            "No items loaded yet. Add an item or import a sample workbook.", color="secondary"
-        )
-    selected_columns = visible_columns(state) or tuple(column_config.keys())
-    compact_lesson_items = level is not None
-    column_widths = {
-        "item": {"minWidth": 72, "maxWidth": 90},
-        "on_hand": {"minWidth": 120},
-        "on_order": {"minWidth": 124},
-        "backorder": {"minWidth": 128},
-        "pna": {"minWidth": 108},
-        "op": {"minWidth": 96},
-        "lp": {"minWidth": 96},
-        "oq": {"minWidth": 96},
-        "eoq": {"minWidth": 96},
-        "soq": {"minWidth": 96},
-        "usage_rate": {"minWidth": 112},
-        "hits_per_month": {"minWidth": 96},
-        "item_cost": {"minWidth": 96},
-        "lead_time": {"minWidth": 120},
-        "standard_pack": {"minWidth": 112},
-        "safety_allowance": {"minWidth": 118},
-        "cp": {"minWidth": 96},
-        "surplus_line": {"minWidth": 168},
-        "days_to_op": {"minWidth": 132},
-        "daily_usage": {"minWidth": 128},
-    }
-    column_defs = []
-    for key in selected_columns:
-        field, header = column_config[key]
-        column_def: dict[str, object] = {"field": field, "headerName": header}
-        if key == "soq":
-            column_def["headerTooltip"] = "Suggested order quantity."
-        if key == "item":
-            column_def["pinned"] = "left"
-        else:
-            column_def["type"] = "numericColumn"
-        column_def.update(column_widths.get(key, {}))
-        column_defs.append(column_def)
-    default_col_def = {"sortable": True, "filter": True, "resizable": True}
-    dash_grid_options = {
-        "pagination": True,
-        "paginationPageSize": min(12, max(1, len(rows))),
-        "paginationPageSizeSelector": False,
-        "animateRows": False,
-    }
-    grid_style = {"height": _workspace_grid_height(state, surface="inventory"), "width": "100%"}
-    if compact_lesson_items:
-        default_col_def.update({"sortable": False, "filter": False})
-        dash_grid_options = {
-            "animateRows": False,
-            "domLayout": "autoHeight",
-            "pagination": False,
-        }
-        grid_style = {"height": "auto", "minHeight": "7.5rem", "width": "100%"}
-    return dag.AgGrid(
-        id="inventory-table-grid",
-        rowData=rows,
-        columnDefs=column_defs,
-        defaultColDef=default_col_def,
-        className=_grid_theme_class(theme),
-        columnSize="sizeToFit",
-        dashGridOptions=dash_grid_options,
-        style=grid_style,
-    )
-
-
-def build_exception_center(state: SimulationState):
-    if not state.exception_center:
-        return dbc.Alert("No ASQ exceptions recorded.", color="secondary")
-    cards = []
-    for exception in list(reversed(state.exception_center[-6:])):
-        cards.append(
-            dbc.Card(
-                dbc.CardBody(
-                    [
-                        html.Div(
-                            [
-                                html.Strong(exception.code),
-                                html.Span(f" • Day {exception.day}", className="ms-2 text-muted"),
-                            ]
-                        ),
-                        html.Div(exception.message, className="mt-2"),
-                        html.Small(
-                            (
-                                f"Item {exception.item_index + 1} • "
-                                f"OP {exception.op:.2f} • ASQ {exception.asq:.2f}"
-                            ),
-                            className="text-muted",
-                        ),
-                    ]
-                ),
-                className="exception-card",
-            )
-        )
-    return html.Div(cards, className="exception-stack")
-
-
-def _reference_definition_block(term: str, definition: str) -> html.Div:
-    return html.Div(
-        [
-            html.Div(term, className="lesson-snapshot-label"),
-            html.Div(definition, className="helper-copy mb-0"),
-        ],
-        className="reference-entry",
-    )
-
-
-def reference_modal_children() -> list:
-    glossary = html.Div(
-        [
-            html.Div("Glossary", className="panel-label"),
-            html.H3("Simulator terms", className="panel-title-small"),
-            html.Div(
-                [
-                    _reference_definition_block(term, definition)
-                    for term, definition in _GLOSSARY_ENTRIES
-                ],
-                className="reference-entry-list",
-            ),
-        ],
-        className="reference-panel",
-    )
-    csd_mapping = html.Div(
-        [
-            html.Div("CSD mapping", className="panel-label"),
-            html.H3("General context", className="panel-title-small"),
-            html.Div(
-                (
-                    "IMSim is an independent inventory management training project. It is "
-                    "not a CSD screen, an Infor tool, or official ERP training."
-                ),
-                className="helper-copy mb-3",
-            ),
-            html.Ul(
-                [
-                    html.Li(
-                        "CSD users may recognize concepts from Product Warehouse Product "
-                        "Setup ordering controls, Demand Center/RRAR workflows, Product "
-                        "Inquiry - Replenishment, and surplus reporting."
-                    ),
-                    html.Li(
-                        "Exact behavior depends on company setup, product line setup, "
-                        "order method, warehouse and product controls, security, and "
-                        "customizations."
-                    ),
-                    html.Li(
-                        "The simulator keeps its SOQ and exception logic intentionally "
-                        "simple so the lessons stay teachable and comparable."
-                    ),
-                ],
-                className="lesson-copy-list",
-            ),
-        ],
-        className="reference-panel",
-    )
-    return [
-        dbc.Tabs(
-            [
-                dbc.Tab(glossary, label="Glossary", tab_id="glossary"),
-                dbc.Tab(csd_mapping, label="CSD Mapping", tab_id="csd-mapping"),
-            ],
-            id="reference-tabs",
-            active_tab="glossary",
-            class_name="reference-tabs",
-        )
-    ]
-
-
-def github_footer_card(github_url: str) -> html.Div:
-    return html.Div(
-        dbc.Card(
-            dbc.CardBody(
-                [
-                    html.Div("IMSim", className="footer-title"),
-                    html.Small(
-                        [
-                            "Project source on ",
-                            html.A(
-                                "GitHub",
-                                href=github_url,
-                                target="_blank",
-                                rel="noopener noreferrer",
-                            ),
-                        ]
-                    ),
-                    html.Small(
-                        "Independent open-source inventory management training project.",
-                        className="d-block mt-2",
-                    ),
-                    html.Small(
-                        ("Not affiliated with, endorsed by, sponsored by, or supported by Infor."),
-                        className="d-block",
-                    ),
-                    " ",
-                    html.Button(
-                        "Hide",
-                        id="gh-footer-hide",
-                        n_clicks=0,
-                        className="imsim-button button-link footer-hide",
-                    ),
-                ]
-            ),
-            className="footer-card",
-        ),
-        id="gh-footer",
-    )
-
-
-def academy_progress_children(state: SimulationState) -> list:
-    total_levels = len(academy_levels())
-    completed = len(state.training.completed_levels)
-    simulator = "Unlocked" if state.training.simulator_unlocked else "Locked"
-    return [
-        dbc.Row(
-            [
-                dbc.Col(
-                    _kpi_card(
-                        "Lessons Complete",
-                        f"{completed}/{total_levels}",
-                        "neutral",
-                        "Academy progress",
-                    ),
-                    md=4,
-                ),
-                dbc.Col(
-                    _kpi_card(
-                        "Next Unlock",
-                        f"Level {min(total_levels, state.training.highest_unlocked_level)}",
-                        "good" if completed else "neutral",
-                        "Highest playable lesson",
-                    ),
-                    md=4,
-                ),
-                dbc.Col(
-                    _kpi_card(
-                        "Simulator",
-                        simulator,
-                        "good" if state.training.simulator_unlocked else "warn",
-                        "Free-play sandbox",
-                    ),
-                    md=4,
-                ),
-            ],
-            className="g-3",
-        )
-    ]
-
-
-def academy_result_children(state: SimulationState):
-    if not state.training.last_result_title and not state.training.last_result_message:
-        return html.Div()
-    tone = "success" if state.training.lesson_status == "passed" else "warning"
-    return dbc.Alert(
-        [
-            html.Strong(state.training.last_result_title or "Academy update"),
-            html.Div(state.training.last_result_message),
-        ],
-        color=tone,
-        class_name="academy-result-alert",
-    )
-
-
-def lesson_tutorial_children(state: SimulationState) -> list:
-    level = active_level(state)
-    if level is None:
-        return [dbc.Alert("Select a lesson from the academy menu.", color="secondary")]
-    framing = []
-    day_basis_helper = [_lesson_day_basis_helper(state)] if _lesson_uses_day_basis(level) else []
-    secondary_notes = _lesson_secondary_notes(level)
-    if level.teaching_goal:
-        framing.append(html.Div(level.teaching_goal, className="helper-copy mb-2"))
-    if level.concept_tags:
-        framing.append(
-            html.Div(
-                [
-                    dbc.Badge(tag, color="secondary", pill=True, class_name="me-1")
-                    for tag in level.concept_tags
-                ],
-                className="mb-2",
-            )
-        )
-    if level.index == 2 and state.items:
-        item = state.items[0]
-        on_order = item_on_order(item)
-        reserved = 0.0
-        committed = 0.0
-        received = 0.0
-        return [
-            *framing,
-            html.Div(level.formula, className="lesson-formula-chip"),
-            html.Div(
-                (
-                    "This lesson keeps Reserved, Committed, and Received at 0 so the "
-                    "purchasing view stays simple."
-                ),
-                className="helper-copy mb-2",
-            ),
-            html.Div(
-                (
-                    "Current PNA = "
-                    f"{item.on_hand:.1f} - {reserved:.1f} - {committed:.1f} - "
-                    f"{item.backorder:.1f} + {on_order:.1f} + {received:.1f} = {item.pna:.1f}"
-                ),
-                className="lesson-formula-chip",
-            ),
-            html.Ul([html.Li(step) for step in level.tutorial_steps], className="lesson-copy-list"),
-            *secondary_notes,
-        ]
-    if level.index == 18 and state.items:
-        return [
-            *framing,
-            html.Div(level.formula, className="lesson-formula-chip"),
-            _emergency_replenishment_panel(state, level),
-            html.Ul([html.Li(step) for step in level.tutorial_steps], className="lesson-copy-list"),
-            *secondary_notes,
-        ]
-    return [
-        *framing,
-        html.Div(level.formula, className="lesson-formula-chip"),
-        *day_basis_helper,
-        html.Ul([html.Li(step) for step in level.tutorial_steps], className="lesson-copy-list"),
-        *secondary_notes,
-    ]
-
-
-def lesson_objective_children(state: SimulationState) -> list:
-    level = active_level(state)
-    if level is None:
-        return [dbc.Alert("No active lesson.", color="secondary")]
-    evaluation = evaluate_active_lesson(state)
-    headline = f"{lesson_days_remaining(state)} day(s) remaining"
-    if evaluation is not None and evaluation.completed:
-        headline = "Lesson window closed"
-    rows = list(evaluation.metric_rows if evaluation is not None else ())
-    if "fill_rate_min" in level.win_conditions:
-        fill = fill_rate(state)
-        fill_value = "n/a" if fill is None else f"{fill * 100:.1f}%"
-        rows.insert(0, f"Current fill rate: {fill_value}")
-    children: list = []
-    if state.training.lesson_status in {"passed", "failed"} and (
-        state.training.last_result_title or state.training.last_result_message
-    ):
-        children.append(
-            dbc.Alert(
-                [
-                    html.Strong(state.training.last_result_title or "Lesson update"),
-                    html.Div(state.training.last_result_message),
-                ],
-                color="success" if state.training.lesson_status == "passed" else "warning",
-                class_name="academy-result-alert",
-            )
-        )
-    children.extend(
-        [
-            html.Div(headline, className="lesson-objective-headline"),
-            html.Ul([html.Li(row) for row in rows], className="lesson-copy-list"),
-        ]
-    )
-    return children
-
-
-def lesson_locked_children(state: SimulationState) -> list:
-    level = active_level(state)
-    if level is None:
-        return [dbc.Alert("No active lesson.", color="secondary")]
-    return [
-        html.Ul(
-            [html.Li(row) for row in level.locked_features],
-            className="lesson-copy-list lesson-locked-list",
-        )
-    ]
-
-
-def lesson_compact_summary_children(state: SimulationState) -> list:
-    level = active_level(state)
-    if level is None:
-        return []
-    evaluation = evaluate_active_lesson(state)
-    headline = (
-        "Lesson window closed"
-        if evaluation is not None and evaluation.completed
-        else f"{lesson_days_remaining(state)} day(s) remaining"
-    )
-    objective_rows = [
-        _compact_lesson_pill_text(row)
-        for row in list(evaluation.metric_rows if evaluation is not None else ())
-    ]
-    return [
-        html.Div(level.formula, className="lesson-compact-chip"),
-        html.Div(
-            [
-                html.Span(headline, className="lesson-compact-pill lesson-compact-pill-strong"),
-                *[html.Span(row, className="lesson-compact-pill") for row in objective_rows],
-            ],
-            className="lesson-compact-inline",
-        ),
-    ]
-
-
-def academy_level_card_children(level_index: int, state: SimulationState) -> list:
-    level = academy_levels()[level_index - 1]
-    status = academy_level_status(state.training, level)
-    status_label = {
-        "completed": "Completed",
-        "unlocked": "Unlocked",
-        "locked": "Locked",
-    }[status]
-    button_label = "Replay Lesson" if status == "completed" else "Start Lesson"
-    return [
-        html.Div(f"Level {level.index}", className="academy-card-kicker"),
-        html.Div(level.title, className="academy-card-title"),
-        html.P(level.summary, className="academy-card-copy"),
-        dbc.Badge(
-            status_label,
-            color="success"
-            if status == "completed"
-            else ("primary" if status == "unlocked" else "secondary"),
-            pill=True,
-            class_name="academy-status-badge",
-        ),
-        html.Div(level.formula, className="academy-card-formula"),
-        html.Button(
-            button_label,
-            id=f"academy-level-{level.index}-button",
-            n_clicks=0,
-            className="imsim-button button-primary button-block mt-3",
-            disabled=status == "locked",
-        ),
-    ]
-
-
-def simulator_unlock_children(state: SimulationState) -> list:
-    unlocked = state.training.simulator_unlocked
-    return [
-        html.Div("Simulator Mode", className="academy-card-title"),
-        html.P(
-            "The full IM dashboard with imports, ASQ, and the sandbox reward controls."
-            if unlocked
-            else "Pass certification to unlock the full simulator and the sandbox reward controls.",
-            className="academy-card-copy",
-        ),
-        dbc.Badge(
-            "Unlocked" if unlocked else "Locked",
-            color="success" if unlocked else "secondary",
-            pill=True,
-            class_name="academy-status-badge",
-        ),
-    ]
+__all__ = [
+    "_plot_base_layout",
+    "_plot_line",
+    "_plot_marker",
+    "_plot_marker_outline",
+    "_grid_theme_class",
+    "academy_level_card_children",
+    "academy_progress_children",
+    "academy_result_children",
+    "build_custom_order_grid",
+    "build_custom_order_rows",
+    "build_exception_center",
+    "build_inventory_figure",
+    "build_inventory_table",
+    "build_kpi_strip",
+    "build_po_overview_grid",
+    "build_po_overview_rows",
+    "costs_card_children",
+    "custom_order_grid_options",
+    "github_footer_card",
+    "inventory_graph_style",
+    "lesson_compact_summary_children",
+    "lesson_locked_children",
+    "lesson_objective_children",
+    "lesson_tutorial_children",
+    "po_overview_grid_options",
+    "reference_modal_children",
+    "refresh_inventory_figure",
+    "sales_card_children",
+    "service_card_children",
+    "simulator_unlock_children",
+]
